app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid>=2.15.2,<3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dill==0.3.8
Django==5.1.1
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==6.1.0
flake8-django==1.4
flake8-import-order==0.18.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
lazy-object-proxy==1.10.0
lxml==5.3.0
mccabe==0.7.0
msgpack==1.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tomlkit==0.13.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zope.interface==7.0.3
